backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy[asyncio]==2.0.35
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
aiosqlite==0.20.0
asyncpg==0.29.0
alembic==1.13.2
redis[asyncio]==5.1.1
bcrypt==4.2.0
pyjwt==2.9.0
python-multipart==0.0.9
sentence-transformers==3.1.1
chromadb==0.5.18
beautifulsoup4==4.12.3
python-docx==1.1.2
pdfplumber==0.11.4
